week05/day5/Mini-Project/bike_store/requirements.txt
asgiref==3.6.0
Django==4.2
django-phonenumber-field==7.1.0
psycopg2==2.9.6
sqlparse==0.4.4
